float instead of int

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
-income = int(input("Enter your monthly income: "))
-expenses = int(input("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+expenses = float(input("Enter your total monthly expenses: "))
 savings = income - expenses
 projected_savings = savings * 12 + (savings * 12 * 0.05)
 print("Your monthly savings are $",savings)
-print("Projected savings after one year, with interest, is: $",int(projected_savings))
+print("Projected savings after one year, with interest, is: $",projected_savings)
